Calculate bivariate table from a vector and a response variable

Calculate bivariate table from a vector and a response variable

bivariate_table(x, target)

Arguments

x
A variable.

target
A binary variable same length as x.

Examples

data(german_credit)

bivariate_table(german_credit$purpose, german_credit$good_bad)

#> # A tibble: 10 x 11
#>    x                 n percent target_n target_rate target_percent non_target_n
#>    <chr>         <int>   <dbl>    <dbl>       <dbl>          <dbl>        <dbl>
#>  1 business         97 0.0970     63.0        0.649         0.0900        34.0 
#>  2 car_new         234 0.234     145          0.620         0.207         89.0 
#>  3 car_used        103 0.103      86.0        0.835         0.123         17.0 
#>  4 domestic_app~    12 0.0120      8.00       0.667         0.0114         4.00
#>  5 education        50 0.0500     28.0        0.560         0.0400        22.0 
#>  6 furniture_or~   181 0.181     123          0.680         0.176         58.0 
#>  7 others           12 0.0120      7.00       0.583         0.0100         5.00
#>  8 radio_or_tel~   280 0.280     218          0.779         0.311         62.0 
#>  9 repairs          22 0.0220     14.0        0.636         0.0200         8.00
#> 10 retraining        9 0.00900     8.00       0.889         0.0114         1.00
#> # ... with 4 more variables: non_target_percent <dbl>, odds <dbl>, woe <dbl>,
#> #   iv <dbl>

bivariate_table(
  cut(german_credit$duration_in_month, c(0, 6, 12, 24, 36, Inf)),
  german_credit$good_bad
)

#> # A tibble: 5 x 11
#>   x            n percent target_n target_rate target_percent non_target_n
#>   <fct>    <int>   <dbl>    <dbl>       <dbl>          <dbl>        <dbl>
#> 1 (0,6]       82  0.0820     73.0       0.890         0.104          9.00
#> 2 (6,12]     277  0.277     210         0.758         0.300         67.0 
#> 3 (12,24]    411  0.411     289         0.703         0.413        122   
#> 4 (24,36]    143  0.143      86.0       0.601         0.123         57.0 
#> 5 (36,Inf]    87  0.0870     42.0       0.483         0.0600        45.0 
#> # ... with 4 more variables: non_target_percent <dbl>, odds <dbl>, woe <dbl>,
#> #   iv <dbl>